feat(io): stop narrowing imported coordinates to float (REQ-101 ±0.002 ft, ADR-054 Phase B, #441)
TASK-228 Phase B. DXF and DWG-native import no longer quantize coordinates to
`float` on the way into the (now `double`) geometry stores.

- LibreDwgCad.cpp: LocalLine / LocalCircle / LocalArc / LocalPolyline / LocalText
  and the ELLIPSE import drop their `static_cast<float>(x - worldDocumentOriginX)`.
- DxfIo.cpp: line / circle / arc / ellipse / polyline / filled-region import keep
  full `double` precision.

CadArc::startRad/sweepRad and CadEllipse::majVx/majVy/ratio revert to `float` —
they are angles / a direction / a ratio, not coordinates (an arc endpoint's error
is r·Δθ, well inside ±0.002 ft at any realistic radius), and `float` there keeps
the DXF writer's byte-stable angle round-trip (issue #111).

DXF writer:
- `worldX` / `worldY` take `double` (were `float` only because the store was).
- The $EXTMIN/$EXTMAX extent sweep snaps every coordinate to the written
  six-decimal grid before accumulating (`q6` / `q6lx` / `q6ly`), so it sees the
  values a reader reconstructs — not the full-precision doubles in memory. With
  `float` storage this was free (a float carries ~6-7 significant digits at model
  magnitude, ≈ the six-decimal text); `double` carries 15+, so a shallow
  huge-radius arc whose centre and radius nearly cancel lands ~1e-6 off on the
  second export without the snap (regression-111, regression-113).

ADR-054 (a) amended: paper-space stores stay `float` — sheet coordinates in paper
inches, where `float` resolves ~1e-6 in, orders of magnitude inside REQ-101's
±0.002 ft. SurveyPoint::easting/northing/elevation is a real `double` target but a
separate, larger change — split to #447 (Phase F).

Build clean; ctest 1359/1359.

+++ src/commands/CadEntities.hpp
@@ -446,8 +446,11 @@ struct CadArc {
   double cx = 0.0;
   double cy = 0.0;
   double r = 0.0;
-  double startRad = 0.0;
-  double sweepRad = 0.0;
+  /// Angles stay `float` (REQ-101 governs *coordinates*; an arc's endpoint error is r·Δθ, and a
+  /// `float` radian ≈ 1e-7 rad keeps that well inside ±0.002 ft at any realistic radius). Keeping
+  /// them `float` also preserves the DXF writer's byte-stable angle round-trip (issue #111).
+  float startRad = 0.f;
+  float sweepRad = 0.f;
   /// Elevation of the arc plane (REQ-057 / ADR-025). Absolute (ADR-025 D2), always 0 in paper
   /// space (ADR-025 (g)). On a tilted arc this is the CENTRE point of the plane, not an
   /// elevation every point on the arc shares.
@@ -724,9 +727,11 @@ inline void CurveEndpointsWorld(const CadArc& a, ray3d::Vec3* outStart, ray3d::V
 struct CadEllipse {
   double cx = 0.0;
   double cy = 0.0;
-  double majVx = 1.0;
-  double majVy = 0.0;
-  double ratio = 0.5;
+  /// Major-axis vector and ratio stay `float` — a direction and a shape ratio, not coordinates
+  /// (see \ref CadArc angle note). Keeps the DXF ellipse round-trip byte-stable (issue #113).
+  float majVx = 1.f;
+  float majVy = 0.f;
+  float ratio = 0.5f;
   /// Elevation of the ellipse's plane (REQ-057 / ADR-025) — parallel to XY, absolute
   /// (ADR-025 D2), always 0 in paper space (ADR-025 (g)). Same rationale as \ref CadArc::z.
   double z = 0.0;
